# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rel-Mon Public API",
    "description": "Public REST API for managing releases, sprints, and dependencies. Authenticated via team-scoped API keys with rate limiting (100 requests/minute per key).",
    "version": "1.0.0",
    "contact": {
      "name": "Rel-Mon Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.rel-mon.com/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
    "/releases": {
      "get": {
        "tags": ["Releases"],
        "summary": "List all releases",
        "description": "Retrieve a list of all releases for your team with optional filtering",
        "operationId": "listReleases",
        "parameters": [
          {
            "name": "sprintId",
            "in": "query",
            "description": "Filter by sprint ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "in": "query",
            "description": "Filter by service ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by single status",
            "schema": {
              "$ref": "#/components/schemas/ReleaseStatus"
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "description": "Filter by multiple statuses (comma-separated)",
            "schema": {
              "type": "string",
              "example": "PLANNING,IN_DEVELOPMENT"
            }
          },
          {
            "name": "ownerId",
            "in": "query",
            "description": "Filter by owner user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isBlocked",
            "in": "query",
            "description": "Filter by blocked status",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page (default: 50, max: 100)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of releases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Release"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer",
                          "description": "Current page number"
                        },
                        "limit": {
                          "type": "integer",
                          "description": "Items per page"
                        },
                        "totalCount": {
                          "type": "integer",
                          "description": "Total number of items"
                        },
                        "totalPages": {
                          "type": "integer",
                          "description": "Total number of pages"
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "description": "Whether there is a next page"
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "description": "Whether there is a previous page"
                        }
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Request limit per window"
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Remaining requests in current window"
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp when the rate limit resets"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": ["Releases"],
        "summary": "Create a new release",
        "description": "Create a new release for your team",
        "operationId": "createRelease",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReleaseInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Release created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Release"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Service or sprint not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Create API keys in the team settings."
      }
    },
    "schemas": {
      "ReleaseStatus": {
        "type": "string",
        "enum": [
          "PLANNING",
          "IN_DEVELOPMENT",
          "IN_REVIEW",
          "READY_STAGING",
          "IN_STAGING",
          "STAGING_VERIFIED",
          "READY_PRODUCTION",
          "DEPLOYED",
          "CANCELLED",
          "ROLLED_BACK"
        ],
        "description": "Release status in the deployment pipeline"
      },
      "SprintStatus": {
        "type": "string",
        "enum": [
          "PLANNING",
          "ACTIVE",
          "COMPLETED",
          "CANCELLED"
        ],
        "description": "Sprint status"
      },
...
      "CreateReleaseInput": {
        "type": "object",
        "required": ["title", "serviceId"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Release title",
            "example": "User Authentication Feature"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "description": "Detailed description of the release",
            "example": "Implement OAuth 2.0 authentication with Google and GitHub providers"
          },
          "serviceId": {
            "type": "string",
            "description": "ID of the service this release belongs to",
            "example": "clx1234567890"
          },
          "sprintId": {
            "type": "string",
            "description": "Optional sprint ID to associate this release with",
            "example": "clx0987654321"
          },
          "version": {
            "type": "string",
            "maxLength": 50,
            "description": "Version number or identifier",
            "example": "v2.3.0"
          },
          "targetDate": {
            "type": "string",
            "format": "date-time",
            "description": "Target deployment date (ISO 8601 format or YYYY-MM-DD)",
            "example": "2026-02-15T00:00:00Z"
          },
          "ownerId": {
            "type": "string",
            "description": "Optional user ID for the release owner",
            "example": "clx1122334455"
          }
        }
      },
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1234567890"
          },
          "name": {
            "type": "string",
            "example": "API Service"
          },
          "color": {
            "type": "string",
            "example": "#3B82F6"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1122334455"
          },
          "firstName": {
            "type": "string",
            "nullable": true,
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "nullable": true,
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://example.com/avatar.jpg"
          }
        }
      },
      "Sprint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx0987654321"
          },
          "name": {
            "type": "string",
            "example": "Sprint 42"
          },
          "status": {
            "$ref": "#/components/schemas/SprintStatus"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-01T00:00:00Z"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-14T23:59:59Z"
          },
          "_count": {
            "type": "object",
            "properties": {
              "releases": {
                "type": "integer",
                "example": 8
              }
            }
          }
        }
      },
...
      "Release": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1234567890"
          },
          "title": {
            "type": "string",
            "example": "User Authentication Feature"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Implement OAuth 2.0 authentication"
          },
          "version": {
            "type": "string",
            "nullable": true,
            "example": "v2.3.0"
          },
          "status": {
            "$ref": "#/components/schemas/ReleaseStatus"
          },
          "isBlocked": {
            "type": "boolean",
            "example": false
          },
          "blockedReason": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "targetDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-02-15T00:00:00Z"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-10T08:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-15T14:22:00Z"
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          },
          "sprint": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Sprint"
              },
              {
                "nullable": true
              }
            ]
          },
          "owner": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              },
              {
                "nullable": true
              }
            ]
          },
          "_count": {
            "type": "object",
            "properties": {
              "items": {
                "type": "integer",
                "example": 5
              },
              "dependsOn": {
                "type": "integer",
                "example": 2
              },
              "dependents": {
                "type": "integer",
                "example": 1
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Resource not found"
          },
          "details": {
            "type": "object",
            "description": "Additional error details (for validation errors)",
            "additionalProperties": true
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication failed - invalid or missing API key",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "Authentication scheme"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Invalid or missing API key"
            }
          }
        }
      },
      "RateLimitError": {
        "description": "Rate limit exceeded (100 requests per minute)",
        "headers": {
          "X-RateLimit-Limit": {
            "schema": {
              "type": "integer"
            },
            "description": "Request limit per window"
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            },
            "description": "Remaining requests (will be 0)"
          },
          "X-RateLimit-Reset": {
            "schema": {
              "type": "integer"
            },
            "description": "Unix timestamp when rate limit resets"
          },
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds until rate limit resets"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Rate limit exceeded. Please try again later."
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Invalid request data",
              "details": {
                "title": "Title is required"
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  }
}